+++ sdks/flutter/lib/synheart_emotion.dart
@@ -61,8 +61,6 @@ export 'src/emotion_error.dart';
 export 'src/features.dart';
 
 // Model inference
-export 'src/model_linear_svm.dart';
-export 'src/json_linear_model.dart';
 export 'src/onnx_model.dart';
 
 // Stream helpers